add listagem e subscription

+++ src/app/controllers/MeetupController.js
@@ -8,25 +8,60 @@ import {
 } from 'date-fns';
 import { Op } from 'sequelize';
 import Meetup from '../models/Meetup';
+import User from '../models/User';
+import File from '../models/File';
+import Subscription from '../models/Subscription';
 
 class MeetupController {
-    async index(req, res) {
-        // retorna todos os meetups de um usuario
-        const meetups = await Meetup.findAll({
-            where: { user_id: req.userId },
-        });
 
-        return res.json(meetups);
+    async index(req, res) {
+
+        const meetups = await Subscription.findAll({
+            where: {
+                user_id: req.userId,
+                // date: {
+                //     [Op.startsWith]: new Date()
+                // }
+            },
+            include: [{
+                model: Meetup,
+                attributes: ['id', 'title', 'date'],
+                where: {
+                    date: {
+                    }
+                }
+            }]
+        })
+
+        return res.json(meetups)
     }
 
     // criar listagem de meetups por data
     async list(req, res) {
         const date = parseISO(req.query.date);
+        const { page } = req.query;
 
         const meetups = await Meetup.findAll({
             where: {
-                [Op.between]: [startOfDay(date), endOfDay(date)],
+                date: {
+                    [Op.between]: [startOfDay(date), endOfDay(date)],
+                }
             },
+            limit: 10,
+            offset: (page - 1) * 10,
+            attributes: ['id', 'title', 'desc', 'locate', 'date'],
+            include: [
+                {
+                    model: User,
+                    as: 'user',
+                    attributes: ['id', 'name', 'email']
+                },
+                {
+                    model: File,
+                    as: 'banner',
+                    attributes: ['path', 'url']
+                },
+            ]
         });
 
         return res.json(meetups);
@@ -63,6 +98,14 @@ class MeetupController {
             return res.status(400).json({
                 error: 'Error in create Meetup: Past dates are not permitted',
             });
+        }
+
+        const findMeetup = await Meetup.findAll({
+            where: { date: hourStart, user_id: req.userId }
+        });
+        console.log(findMeetup)
+        if (findMeetup.length != 0) {
+            return res.status(400).json({ error: 'You must have a meetup on this date' })
         }
 
         const meetup = await Meetup.create({
@@ -113,7 +156,15 @@ class MeetupController {
     }
 
     async delete(req, res) {
-        await Meetup.destroy({ where: { id: req.body.id } });
+
+        const meetup = await Meetup.findByPk(req.body.id);
+        if (meetup.user_id !== req.userId) {
+            return res.status(401).json({
+                error: 'You don´t have permissions to delete this meetup, only owners can do'
+            })
+        }
+
+        await Meetup.destroy({ where: { id: req.body.id, user_id: req.userId } });
 
         const meetups = await Meetup.findAll({
             where: { user_id: req.userId },
